+++ tests/Strategy/AuthorizationHeaderTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace ApiAuth\Test\Strategy;
 
 use ApiAuth\AuthData;
-use ApiAuth\Strategy\AuthorizationHeader;
+use ApiAuth\Strategy\BasicAuthorizationHeader;
 use Laminas\Diactoros\ServerRequest;
 use PHPUnit\Framework\TestCase;
 
@@ -14,13 +14,13 @@ use function base64_encode;
 use function sprintf;
 
 /**
- * @covers \ApiAuth\Strategy\AuthorizationHeader
+ * @covers \ApiAuth\Strategy\BasicAuthorizationHeader
  */
 class AuthorizationHeaderTest extends TestCase
 {
     public function testMissingHeader(): void
     {
-        $strategy = new AuthorizationHeader();
+        $strategy = new BasicAuthorizationHeader();
         $request  = new ServerRequest();
         $this->expectExceptionCode(401);
         ($strategy)($request);
@@ -28,14 +28,14 @@ class AuthorizationHeaderTest extends TestCase
 
     public function testMissingHeaderHotRequired(): void
     {
-        $strategy = new AuthorizationHeader(false);
+        $strategy = new BasicAuthorizationHeader(false);
         $request  = new ServerRequest();
         $this->assertNull(($strategy)($request));
     }
 
     public function testEmptyHeader(): void
     {
-        $strategy = new AuthorizationHeader();
+        $strategy = new BasicAuthorizationHeader();
         $request  = (new ServerRequest())->withHeader('Authorization', '');
         $this->expectExceptionCode(401);
         ($strategy)($request);
@@ -43,7 +43,7 @@ class AuthorizationHeaderTest extends TestCase
 
     public function testInvalidHeader(): void
     {
-        $strategy = new AuthorizationHeader();
+        $strategy = new BasicAuthorizationHeader();
         $request  = (new ServerRequest())->withHeader('Authorization', 'Basic: 123');
         $this->expectExceptionCode(401);
         ($strategy)($request);
@@ -54,7 +54,7 @@ class AuthorizationHeaderTest extends TestCase
         $identity   = 'abc';
         $credential = '123';
 
-        $strategy = new AuthorizationHeader();
+        $strategy = new BasicAuthorizationHeader();
         $request  = (new ServerRequest())->withHeader(
             'Authorization',
             'Basic: ' . base64_encode(sprintf('%s:%s', $identity, $credential))